модуляция js ч.3
- скрипты работы модальных окон переработаны и выведен в отдельный файл
- работа протестирована, ошибок нет, все функции активны
- функции переработаны в переменные
